refactor(media): Adopt layered module naming in media domain
