+++ src/domains/media/cache/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @module @domains/media/cache
+ * @remarks Barrel exports for optimized media image cache helpers keyed by semantic paths
+ * or direct source URLs.
+ * @see {@link ./media-cache} for serialization and cache key construction
+ * @example
+ * ```typescript
+ * import { mediaCache } from '@domains/media/cache'
+ * ```
+ */
+export * from './media-cache'
