+++ src/domains/media/schemas/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @module @domains/media/schemas
+ * @remarks Barrel exports for media route validation schemas.
+ * @see {@link ./media-schema} for the media proxy request schema
+ * @example
+ * ```typescript
+ * import { mediaRequestSchema } from '@domains/media/schemas'
+ * ```
+ */
+export * from './media-schema'
