--- src/domains/media/services/image-optimizer.ts
@@ -1,11 +1,0 @@
-import {
-  optimizeImageBuffer,
-  type OptimizeOptions,
-} from '@/core/utils/image/optimize'
-
-export const optimizeMediaImageBuffer = (
-  buffer: Buffer,
-  options: OptimizeOptions = {}
-) => {
-  return optimizeImageBuffer(buffer, options)
-}
