+++ src/domains/media/services/index.ts
@@ -1,0 +1,15 @@
+/**
+ * @module @domains/media/services
+ * @remarks Barrel exports for media application services covering remote fetching, image
+ * optimization, and the full semantic-path delivery pipeline.
+ * @see {@link ./media-service} for orchestration
+ * @see {@link ./media-fetch-service} for HTTP downloads
+ * @see {@link ./image-optimizer-service} for buffer optimization
+ * @example
+ * ```typescript
+ * import { mediaService, fetchMediaAsset, optimizeMediaImageBuffer } from '@domains/media/services'
+ * ```
+ */
+export * from './image-optimizer-service'
+export * from './media-fetch-service'
+export * from './media-service'
