+++ src/domains/media/types/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @module @domains/media/types
+ * @remarks Barrel exports for shared media domain types covering semantic paths, assets,
+ * request shapes, and service configuration.
+ * @see {@link ./media-types} for all type definitions
+ * @example
+ * ```typescript
+ * import type { SemanticMediaPath, MediaAsset, OptimizedMedia } from '@domains/media/types'
+ * ```
+ */
+export * from './media-types'
